### 2018-04-29 #709 @whedon
[PRE REVIEW]: dit: a Python package for discrete information theory
**Submitting author:** @Autoplectic (<a href="http://orcid.org/0000-0003-2149-9085">Ryan G. James</a>)
**Repository:** <a href="https://github.com/dit/dit" target ="_blank">https://github.com/dit/dit</a>
**Version:** v1.2.0
