+++ docs/references.bib
@@ -484,6 +484,18 @@
   year={2017}
 }
 
+% i_rr
+@article{goodwell2017temporal,
+  title={Temporal information partitioning: Characterizing synergy, uniqueness, and redundancy in interacting environmental variables},
+  author={Goodwell, Allison E and Kumar, Praveen},
+  journal={Water Resources Research},
+  volume={53},
+  number={7},
+  pages={5920--5942},
+  year={2017},
+  publisher={Wiley Online Library}
+}
+
 % PED and h_cs
 @article{ince2017partial,
   title={The {P}artial {E}ntropy {D}ecompostion: decomposing multivariate entropy and mutual information via pointwise common surprisal},
