### 2018-05-10 ac9fd4d
update license file a bit for automatic github detection.
