### 2018-05-10 256e840
fix license info in readme
