+++ docs/basicinfo.rst.txt
@@ -45,4 +45,4 @@ Code and bug tracker:
    https://github.com/dit/dit
 
 License:
-   BSD 2-Clause, see LICENSE.txt for details.
+   BSD 3-Clause, see LICENSE.txt for details.
