### 2018-05-14 #709 @Autoplectic
@arfon khinsen, rougier, jmbr, kczimm, stsievert, vsoch would all probably be fine reviewers.
