### 2018-05-15 #709 @arfon
@kczimm, @stsievert - would you be willing to review this submission for JOSS?
