### 2018-05-15 #709 @stsievert
I'd be happy to! I've got a deadline on Friday, but I should be free after.
