### 2018-05-15 #709 @stsievert
Here's a good list of some reviewer guidelines: http://joss.theoj.org/about#reviewer_guidelines
