### 2018-05-16 #709 @arfon
@kczimm, @stsievert - many thanks for agreeing to review this. I'll go ahead and set up the main review issue next.
