### 2018-05-16 #738 @whedon
[REVIEW]: dit: a Python package for discrete information theory
**Submitting author:** @Autoplectic (<a href="http://orcid.org/0000-0003-2149-9085">Ryan G. James</a>)
**Repository:** <a href="https://github.com/dit/dit" target ="_blank">https://github.com/dit/dit</a>
**Version:** v1.2.0

- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? (well, a LICENSE.txt file)
