### 2018-05-16 #709 @arfon
@kczimm, @stsievert, @Autoplectic - please head over to #738 for the actual review.
