### 2018-05-18 #738 @kczimm
@afron I am still testing the functionality of this software package. In the meantime, I've a few questions. I'm not familiar with DOI or how to answer the following question: "Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?" I didn't see anything in the reviewer guidelines that could help my understanding of this question. Also, my minor feedback for the authors at this time is to increase the statement of need and intended audience in their documentation. Is this minor feedback worth creating an issue in their repository? Thanks in advance for the help.
