### 2018-05-18 #738 @stsievert
@Autoplectic I'm guessing @chebee7i is Christopher J. Ellison and James P. Crutchfield is your advisor. Is that correct?
