### 2018-05-18 dit/dit#141 @stsievert
Include statement of need on doc main page
I don't think "`dit` is a Python package for discrete information theory." is enough motivation for this software. I think listing the functions `dit` provides and the problems it can solve.

My favorite examples to point to for motivational documentation are with [distributed][1] and [dask][2]. The list bullet points of the problems their software is designed to solve.

[1]:https://distributed.readthedocs.io/en/latest/
[2]:https://dask.pydata.org/en/latest/docs.html
