### 2018-05-18 e64c010
DOC: make install it's own section
