+++ docs/install.rst
@@ -1,0 +1,18 @@
+Install
+=======
+
+The easiest way to install is:
+
+.. code-block:: bash
+
+  pip install dit
+
+Alternatively, you can clone this repository, move into the newly created
+``dit`` directory, and then install the package:
+
+.. code-block:: bash
+
+  git clone https://github.com/dit/dit.git
+  cd dit
+  pip install .
+
