+++ docs/basicinfo.rst.txt
@@ -22,22 +22,6 @@ Optional Dependencies:
    * `pint <https://pint.readthedocs.io>`_
    * `scikit-learn <http://scikit-learn.org/>`_
 
-Install:
-   The easiest way to install is:
-
-   .. code-block:: bash
-
-      pip install dit
-
-   Alternatively, you can clone this repository, move into the newly created
-   ``dit`` directory, and then install the package:
-
-   .. code-block:: bash
-
-      git clone https://github.com/dit/dit.git
-      cd dit
-      pip install .
-
 Mailing list:
    None
 
