### 2018-05-20 dit/dit#144 @stsievert
Optional dependency clarification on the readme?
There a bunch of dependencies listed on the readme, both required and optional. A couple questions:

* Why are the optional dependencies listed? Maybe make a note like "installing cython can result {benefits}. Run `pip install cython` or see [the Cython install doc page][1] for more detail".
* Why are the dependencies listed? Shouldn't my package manager take care of that? If you want to list them, I'd suggest moving them further down the readme.

I think there should be some more clarification around why these are listed, what advantages they bring and how to install these packages.

Also, it also looks like this list is not maintained. In [requirements_optional.txt][2], there are a couple packages that are not listed in the README or on the docs:

* hypothesis
* ipython
* matplotlib

[2]:https://github.com/dit/dit/blob/256e8409bf7cb96c43acbb651da06a78866643b9/requirements_optional.txt

[1]:http://docs.cython.org/en/latest/src/quickstart/install.html
