### 2018-05-20 b0d8a9f
DOC: detail steps to test dit
