### 2018-05-20 0cd0d70
MAINT: better error message in construct_alphabets
