### 2018-05-21 dit/dit#144 @stsievert
To get around the out-of-date issue, I'd recommend listing the required dependencies in `requirements.txt`. On the docs for the optional requirements, I think it'd be nice to have

1. What features this optional dependency brings
2. How to install
