### 2018-05-21 #738 @stsievert
I've gone through this repo, and have a couple comments. This software looks good and almost ready for acceptance, but I have a couple comments.

* I didn't see a clear statement of need in the paper or in the docs. This should clearly list the problems your software solves (not just the functions it provides) and the target audience. https://github.com/dit/dit/issues/141
* I didn't see any DOIs linked in `paper.bib` though the references were complete (and I would consider suitable for another academic journal). @afron is this acceptable?
* I didn't see any guidelines for users to contribute to your software. I think this is most often in CONTRIBUTING.md, which GitHub has support for.
* The dependency list confused me, especially the optional dependencies. There weren't any reasons why some dependencies were optional, and some of them were out of date. I've filed https://github.com/dit/dit/issues/144 for this.
* I didn't find steps to test the software, thought I figured it out and filed https://github.com/dit/dit/pull/146

I've put in some other issues and PRs too, but these are the most important issues right now.
