### 2018-05-21 #738 @arfon
> I didn't see any DOIs linked in paper.bib though the references were complete (and I would consider suitable for another academic journal). @afron is this acceptable?

@stsievert - references without DOIs are acceptable provided they _don't actually have one_, i.e. a journal publications such as [this one](https://github.com/dit/dit/blob/master/joss_paper/paper.bib#L93) is missing this DOI: https://doi.org/10.1016/j.physleta.2011.10.066

@Autoplectic - please go through your `paper.bib` and add a DOI for each entry where you can find a DOI.
