### 2018-05-21 d8136ac
DOC: make install it's own section (#142)
