+++ docs/generalinfo.rst
@@ -4,3 +4,5 @@ General Information
 .. include:: basicinfo.rst.txt
 
 .. include:: quickstart.rst.txt
+
+.. include:: install.rst.txt
