### 2018-05-21 35c3428
MAINT: better error message in construct_alphabets (#147)
* MAINT: better error message in construct_alphabets

* MAINT: point to potential solutions
