### 2018-05-21 e96d11b
DOC: use python 3 print (#143)
