+++ requirements_optional.txt
@@ -2,7 +2,6 @@ colorama
 Cython
 hypothesis >= 3.16.0
 hypothesis[numpy]
-ipython
 matplotlib
 numdifftools
 pint
