### 2018-05-21 dit/dit#144 @Autoplectic
I've listed the benefit granted by each optional dependency now, and brought the two lists into alignment.
