### 2018-05-21 #738 @Autoplectic
> @Autoplectic I'm guessing @chebee7i is Christopher J. Ellison and James P. Crutchfield is your advisor. Is that correct?

Largely. CJE & I were both graduate students under JPC. CJE is now in industry, and I am back with JPC as a postdoc.
