### 2018-05-21 86e40d2
Add introduction to readme & docs
