### 2018-05-21 acbe2df
adjust sphinx conf
