### 2018-05-21 eeef81e
add DOIs where possible.
