+++ joss_paper/paper.bib
@@ -8,7 +8,8 @@
     pages = {776},
     publisher = {Wiley-Interscience},
     title = {{Elements of Information Theory}},
-    year = {2006}
+    year = {2006},
+    doi = {10.1002/047174882X}
 }
 
 @book{MacKay2003,
@@ -22,7 +23,8 @@
     title={Information theory and network coding},
     author={Yeung, R. W.},
     year={2008},
-    publisher={Springer}
+    publisher={Springer},
+    doi = {10.1007/978-0-387-79234-7}
 }
 
 @url{dit,
@@ -53,7 +55,8 @@
   number={2},
   pages={97--116},
   year={1954},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1109/TIT.1954.1057469}
 }
 
 % total correlation
@@ -65,7 +68,8 @@
   number={1},
   pages={66--82},
   year={1960},
-  publisher={IBM}
+  publisher={IBM},
+  doi={10.1147/rd.41.0066}
 }
 
 % dual total correlation
@@ -76,7 +80,8 @@
   volume={46},
   pages={26--45},
   year={1980},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0019-9958(80)90478-7}
 }
 
 @article{Han1975linear,
@@ -86,7 +91,8 @@
   volume={29},
   pages={337--368},
   year={1975},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0019-9958(75)80004-0}
 }
 
 @article{Abdallah2012,
@@ -97,7 +103,8 @@
   number={4},
   pages={275--281},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.physleta.2011.10.066}
 }
 
 % CAEKL mutual information
@@ -109,7 +116,8 @@
   number={10},
   pages={1883--1913},
   year={2015},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/JPROC.2015.2458316}
 }
 
 %% common informations
@@ -134,7 +142,8 @@
   number={10},
   pages={6337--6350},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TIT.2011.2165807}
 }
 
 % wyner common information
@@ -146,7 +155,8 @@
   number={2},
   pages={163--179},
   year={1975},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TIT.1975.1055346}
 }
 
 % multivariate wyner common information
@@ -156,7 +166,8 @@
   booktitle={Communication, Control, and Computing (Allerton), 2010 48th Annual Allerton Conference on},
   pages={836--843},
   year={2010},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/ALLERTON.2010.5706995}
 }
 
 % exact common information
@@ -166,15 +177,21 @@
   booktitle={Information Theory (ISIT), 2014 IEEE International Symposium on},
   pages={161--165},
   year={2014},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/ISIT.2014.6874815}
 }
 
 % marginal utility of information
-@article{Allen2014,
-  title={An information-theoretic formalism for multiscale structure in complex systems},
-  author={Allen, B. and Stacey, B. C. and Bar-Yam, Y.},
-  journal={arXiv preprint arXiv:1409.4708},
-  year={2014}
+@article{allen2017multiscale,
+  title={Multiscale information theory and the marginal utility of information},
[…]
