### 2018-05-21 #738 @Autoplectic
I have added DOIs where possible.
