### 2018-05-21 #738 @Autoplectic
I believe I have addressed all the reviewers concerns at this point.
