### 2018-05-21 #738 @Autoplectic
scratch that --- working on an issue with the docs. the custom mathjax works fine under docs.dit.io, but fails under dit.readthedocs.io
