### 2018-05-21 9c22791
more latex fixes
