### 2018-05-21 #738 @Autoplectic
OK, the docs work fine at http://dit.readthedocs.io and at http://docs.dit.io, which is the official link. I'm not sure why it doesn't work at https://dit.readthedocs.io. Since it works at the official link, which is the only one in the readme and such, things should be ok.
