### 2018-05-21 #738 @stsievert
Thanks @Autoplectic, both for the quick response and handling these issues well.

Maybe I'm missing it, but have you addressed contributing issue mentioned in https://github.com/openjournals/joss-reviews/issues/738#issuecomment-390523494?
