+++ CONTRIBUTING.md
@@ -1,0 +1,4 @@
+Contributing
+------------
+
+If you'd like a feature added to ``dit`` or notice any problems, please file an issue, or, even better, open a pull request. We'll work with you to ensure that the code is tested and documented.
