### 2018-05-22 #738 @Autoplectic
@stsievert The information about contributing is at the bottom of the readme. I've also added a CONTRIBUTING.md as you have suggested.

Thank you @stsievert, @kczimm, and @arfon for your help shepherding this submission through!
