### 2018-05-22 #738 @kczimm
I had many of the same comments as @stsievert. I don't have any other comments or suggestions at this time.
