### 2018-05-22 e2698d4
DOC: add more detail to contributing
