+++ CONTRIBUTING.md
@@ -2,3 +2,9 @@ Contributing
 ------------
 
 If you'd like a feature added to ``dit`` or notice any problems, please file an issue, or, even better, open a pull request. We'll work with you to ensure that the code is tested and documented.
+
+It's appropriate to file an issue when you encounter a problem with `dit`. Be sure to include
+
+* the `dit` version you're using
+* the operating system and Python version
+* full tracebacks
