### 2018-05-22 #738 @stsievert
> I've also added a CONTRIBUTING.md as you have suggested.

Could you add a line for people seeking support? This could be a link to a Gitter channel, a tip to stackoverflow with some tag or file an issue.

 Here's the GitHub support for CONTRIBUTING.md: https://blog.github.com/2012-09-17-contributing-guidelines/

It's easy to check off boxes when everything (or most things) are done right.
