### 2018-05-22 #738 @Autoplectic
@stsievert i've added to the contributing.md file!
