### 2018-05-22 b55e715
DOC: add more detail to contributing (#148)
