+++ dit/__init__.py
@@ -13,7 +13,7 @@ Python package for sigma-algebras defined on finite sets.
 
 """
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"
 
 # Order is important!
 from .params import ditParams
