### 2018-05-30 #738 @arfon
@Autoplectic - I get the following error when compiling the doc:

```
pandoc-citeproc: reference Allen2014 not found
```
Can you check this reference? Is it missing in the bibtex or just cited incorrectly in the `paper.md`
