### 2018-05-31 8b9f993
fix reference in joss paper
