+++ joss_paper/paper.md
@@ -34,7 +34,7 @@ hope that having all these measures and techniques implemented in one place will
 development of robust techniques for the automated quantification of dependencies within a system
 and concrete interpretation of what those dependencies mean.
 
-``dit`` implements the vast majority of information measure defined in the literature, including entropies (Shannon[@Cover2006], Renyi, Tsallis), multivariate mutual informations (co-information[@Bell2003][@mcgill1954multivariate], total correlation[@watanabe1960information], dual total correlation[@te1980multiple][@Han1975linear][@Abdallah2012], CAEKL mutual information[@chan2015multivariate]), common informations (Gács-Körner[@Gacs1973][@tyagi2011function], Wyner[@wyner1975common][@liu2010common], exact[@kumar2014exact], functional, minimal sufficient statistic), and channel capacity[@Cover2006]. It includes methods of studying joint distributions including information diagrams, connected informations[@Schneidman2003][@Amari2001], marginal utility of information[@Allen2014], and the complexity profile[@Baryam2004]. It also includes several more specialized modules including bounds on the secret key agreement rate[@maurer1997intrinsic], partial information decomposition[@williams2010nonnegative], rate-distortion theory[@Cover2006] & information bottleneck[@tishby2000information], and others. Please see the [``dit`` homepage](https://github.com/dit/dit) for a complete and up-to-date list.
+``dit`` implements the vast majority of information measure defined in the literature, including entropies (Shannon[@Cover2006], Renyi, Tsallis), multivariate mutual informations (co-information[@Bell2003][@mcgill1954multivariate], total correlation[@watanabe1960information], dual total correlation[@te1980multiple][@Han1975linear][@Abdallah2012], CAEKL mutual information[@chan2015multivariate]), common informations (Gács-Körner[@Gacs1973][@tyagi2011function], Wyner[@wyner1975common][@liu2010common], exact[@kumar2014exact], functional, minimal sufficient statistic), and channel capacity[@Cover2006]. It includes methods of studying joint distributions including information diagrams, connected informations[@Schneidman2003][@Amari2001], marginal utility of information[@allen2017multiscale], and the complexity profile[@Baryam2004]. It also includes several more specialized modules including bounds on the secret key agreement rate[@maurer1997intrinsic], partial information decomposition[@williams2010nonnegative], rate-distortion theory[@Cover2006] & information bottleneck[@tishby2000information], and others. Please see the [``dit`` homepage](https://github.com/dit/dit) for a complete and up-to-date list.
 
 Where possible, the implementations in ``dit`` support multivariate, conditional forms even if not defined that way in the literature. For example, ``dit`` implements the multivariate, conditional exact common information even though it was only defined for two variables.
 
