### 2018-05-31 #738 @Autoplectic
Fixed. I updated the bib file to reference the published version of that paper, but didn't fix the ref in the paper.
